header file modified

+++ 0x15-file_io/0-read_textfile.c
@@ -11,7 +11,7 @@
  */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	ssize_t o, r, w;
+	int o, r, w;
 	char *buffer;
 
 	if (filename == NULL)
